Replaced DG412s with DG409s
Replaced the DG412s with DG409s. This reduces the leakage current form
the switches to +-1nA per device. Also reduces the number of control
lines to 10. Settled on an overvolt protection scheme. Blocking diodes
on the supply lines for the DG409s, as per the data covering, and rail
clamping diodes with current limiting resistors.

As an overvolt condition starts, the protection diodes in the DG409s
start to conduct at about 0.3v, the current is blocked from the power
rail by the blocking diode, as per the datasheet, the fault voltage
starts to power the device and float it up to the fault voltage. the
current is limited by the device? This continues until the clamping
diodes start to conducted and clamp the fault voltage to the power rail.